+++ XHTD_SERVICES.Data/Entities/tblStoreOrderOperating.cs
@@ -123,7 +123,7 @@ namespace XHTD_SERVICES.Data.Entities
         public Nullable<decimal> ExportedNumber { get; set; }
         public Nullable<decimal> ExtraNumber { get; set; }
         public string ExtraReason { get; set; }
-        public Nullable<double> MachineExportedNumber { get; set; }
+        public Nullable<decimal> MachineExportedNumber { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<tblExportHistory> tblExportHistories { get; set; }
